Enhancement: Move indirect tests into separate namespace
